pyproject.toml
[project]
name = "blockchain-ingestor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "web3>=7.14.1",
]
[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
